pyproject.toml
[tool.poetry]
name = "puls-events-rag"
version = "1.0.0"
description = "RAG for cultural event recommendation"
authors = ["Iyed Jaziri <iyedjaziri@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
faiss-cpu = "^1.7.4"
mistralai = "^1.0.0"
transformers = "^4.36.0"
sentence-transformers = "^2.2.2"
torch = "^2.1.0"

# API Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Data Processing
pandas = "^2.1.0"
numpy = "^1.26.0"
scikit-learn = "^1.3.0"

requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
lxml = "^5.1.0"
tqdm = "^4.66.0"
loguru = "^0.7.2"
datasets = "^2.16.0"
langchain-mistralai = "^0.2.0"
ragas = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
httpx = "^0.28.1"

# Code Quality
black = "^23.12.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.0"
pylint = "^3.0.0"
jupyter = "^1.0.0"
ipykernel = "^6.29.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
plotly = "^5.18.0"

[tool.poetry.group.eval]
optional = true

[tool.poetry.group.eval.dependencies]
# RAG Evaluation
evaluate = "^0.4.1"
nltk = "^3.8.0"
faiss-cpu = "^1.7.4"
datasets = "^2.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310', 'py311']

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = ['tests/', 'notebooks/']

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=scripts --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["scripts"]
omit = ["tests/*", "notebooks/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
